Modification of Epson ET-M1120 printer to eject resin onto substrates in a desired thickness and pattern
This modification of an Epson ET-M1120 printer allows Stensborg's resin to be ejected onto a substrate in a specific thickness and pattern after which Stensborg's desktop machine is ideal for imprinting micro and nano structures. For small batches, prototypes and teaching it substitutes the traditional K-bar while maintaining a nice even layer and allowing specific patterns to be printed if for example only a certain part of a substrate requires an imprinted pattern.